[refactor] user views 기능별로 분리
