+++ users/urls.py
@@ -1,12 +1,17 @@
 from django.urls import path
-from .views import *
+
+from users.views.user_check_views import *
+from users.views.user_diag_record_views import *
+from users.views.user_get_update_delete_views import *
+from users.views.user_login_views import *
+from users.views.user_sign_up_views import SignupView
 
 urlpatterns = [
-    path('', UserListView.as_view()),
+    path('', GetUsersAndUpdateView.as_view()),
     path('/signup', SignupView.as_view()),
     path('/login', LoginView.as_view()),
     path('/<int:userId>/auto/login', AutoLoginView.as_view()),
-    path('/<int:userId>', UserView.as_view()),
+    path('/<int:userId>', GetUserAndDeleteView.as_view()),
     path('/checknickname', CheckNicknameView.as_view()),
     path('/recordsave', RecordSaveView.as_view()),
     path('/prevrecord', GetDiagRecordView.as_view()),
